Added Weather Report Functionality
Added methods to WeatherReport.java that aggregate an array of
StationReport objects to produce an array of WeatherReport objects.
The StationReport objects are binned by N hours where N < 24.
The API allows for the initial beginning to take place from the earliest
StationReport or at midnight.
The StationReport values are averaged for the specified time interval.
An order of preference of stations for filling in a WeatherReport can be
specified. For instance, if DMH does not have values for some variables,
BWI can be used to fill in the missing data (assuming BWI has the data).

The Driver now reads the StationReport data and generates the
WeatherReport array.